Consolidate xorb uploader to session.
